#  [IndVarSimplify] Eliminated Pure LoopCounter