This patch accomplishes two main tasks:

- It relaxes the stride restriction on the LoopCounter to provide additional optimization opportunities; 
- It eliminates the Pure LoopCounter, thereby unlocking potential for further optimizations, such as LoopUnroll.

**Key changes**:

- Relax the stride restriction on the LoopCounter

  Usually the **loop's iterations are counted** by an integer-valued variable that proceeds **upward (or downward) by a constant** amount with each iteration[1][2].  However, the current design requires the loop counter to have a unit stride (+1) and does not support -1, which limits optimization potential.

  Without disrupting the original design, we have relaxed the stride restrictions in both `isLoopCounter()` and `genLoopLimit()` to allow loop counters with strides of either +1 or -1. This enhancement, combined with LLVM’s existing infrastructure, enables support for common countdown loop patterns—for example, transforming a loop such as `for (i = 31; i * i > 48; i--)` into `for (i = 31; i > 6; i--)`—thereby improving the pass’s versatility and optimization coverage.

  

  [1] Steven S. Muchnick. 1998. Advanced compiler design and implementation. Morgan Kaufmann Publishers Inc., San Francisco, CA, USA.

  [2] [Loops - Using and Porting GNU Fortran](https://gcc.gnu.org/onlinedocs/gcc-3.3.6/g77/Loops.html)

- Eliminate the Pure LoopCounter

  We define a PHI node as the Pure LoopCounter if it meets these three conditions: (1) it is used for loop termination testing;  (2) it has at most two users, PostIncOrDec and Cmp (optional). (3) its type is of integer type.

  This optimization is based on a key insight: the Pure LoopCounter can be replaced by another induction variable that satisfies the following conditions: (1) it meets the criteria defined by `isLoopCounter()`; (2) it is of integer type with a width equal to that of `ExitCount`; and (3) it is not the Pure LoopCounter itself.

  Eliminating the Pure LoopCounter can expose additional opportunities for downstream optimizations, such as loop unroll. For example

  ```fortran
    DO j = jms, jme
      DO k = kms, kme
        DO i = ims, ime
          u(i,k,j) = u(i,k,j) + dts * ru_tend(i,k,j)
        ENDDO 
      ENDDO
    ENDDO
  ```

  This common example can successfully perform four loop unrollings after eliminating the pure loop counter, resulting in a **36%** performance improvement.

**Key details**

- Only active when  `-enable-pure-loop-counter-elimination` is enabled(disabled by default)

**Tests**

- Two LIT tests  added to confirm the effectiveness of the patch.

- Four existing LIT tests  modified because relaxing the stride restrictions on LoopCounters provides more optimization opportunities for LFTR, such as simplifying exit comparisons from `sgt/slt` to `eq/ne`.

- We further verified the correctness of the transformation and evaluated its performance using the SPEC CPU2006 and SPEC CPU2017 benchmark suites. 

- The results of the SPEC CPU2017 benchmark are as follows:

  | bench  | ratio<br>[llvm base] | ratio<br>[with the relaxed LoopCounter] |  SpeedUp  | ratio<br>[with this PR] |  SpeedUp  |
  | ------------------------------------------------------------ | :------------------: | :-------------------------------------: | :-------: | :---------------------: | :-------: |
  | 623.xalancbmk_s                                              |        12.86         |                  13.01                  | **1.01x** |          13.01          | **1.01x** |
  | 648.exchange2_s                                              |        17.38         |                  18.87                  | **1.09x** |          18.88          | **1.09x** |
  | 619.lbm_s                                                    |         7.95         |                  8.08                   | **1.02x** |          8.13           | **1.02x** |
  | 621.wrf_s                                                    |        11.04         |                  11.02                  |   1.00x   |          11.19          | **1.01x** |
  | 628.pop2_s                                                   |         3.75         |                  3.75                   |   1.00x   |          3.78           | **1.01x** |
  | 654.roms_s                                                   |        10.74         |                  10.70                  |   1.00x   |          10.95          | **1.02x** |

  * \[with the relaxed LoopCounter]: Relax the stride restriction on the LoopCounter.

  * \[with this PR]: Relax stride constraints and eliminate the Pure LoopCounter.

  * Platform: X86-intel (I9-11900K, L1-cache 384KB, L2-cache 4MB, L3-cache 16MB, cache line 64B).

**Authors**

  The [XSCC compiler team](https://github.com/orgs/OpenXiangShan/teams/xscc) developed this implementation.