@nikic 2025-07-03
Without looking into too much detail, this sounds like what LoopTermFold is doing? cc @preames